Move unexpected module to ctypes
