+++ core/src/consensus/cuckoo/mod.rs
@@ -21,10 +21,10 @@ use std::cmp::{max, min};
 use byteorder::{ByteOrder, LittleEndian};
 use ccrypto::blake256;
 use ctypes::machine::WithBalances;
+use ctypes::util::unexpected::{Mismatch, OutOfBounds};
 use cuckoo::Cuckoo as CuckooVerifier;
 use primitives::U256;
 use rlp::UntrustedRlp;
-use unexpected::{Mismatch, OutOfBounds};
 
 use self::params::CuckooParams;
 use super::super::block::{ExecutedBlock, IsBlock};
