+++ core/src/consensus/tendermint/mod.rs
@@ -26,13 +26,13 @@ use ccrypto::blake256;
 use ckey::{public_to_address, recover, Address, Message, Password, Signature};
 use cnetwork::{Api, NetworkExtension, NodeId, TimerToken};
 use ctypes::machine::WithBalances;
+use ctypes::util::unexpected::{Mismatch, OutOfBounds};
 use ctypes::BlockNumber;
 use parking_lot::{Mutex, RwLock};
 use primitives::{Bytes, H256, U128, U256};
 use rand::{thread_rng, Rng};
 use rlp::{self, Decodable, DecoderError, Encodable, RlpStream, UntrustedRlp};
 use time::Duration;
-use unexpected::{Mismatch, OutOfBounds};
 
 use self::message::*;
 pub use self::params::{TendermintParams, TendermintTimeouts};
