+++ core/src/verification/verification.rs
@@ -17,11 +17,11 @@
 use std::time::{SystemTime, UNIX_EPOCH};
 
 use cmerkle::skewed_merkle_root;
+use ctypes::util::unexpected::{Mismatch, OutOfBounds};
 use ctypes::BlockNumber;
 use heapsize::HeapSizeOf;
 use primitives::{Bytes, H256};
 use rlp::UntrustedRlp;
-use unexpected::{Mismatch, OutOfBounds};
 
 use super::super::blockchain::BlockProvider;
 use super::super::client::{BlockInfo, TransactionInfo};
