+++ network/Cargo.toml
@@ -20,4 +20,3 @@ rlp_derive = { path = "../util/rlp_derive" }
 slab = "0.4"
 table = { path = "../util/table" }
 time = "0.1"
-unexpected = { path = "../util/unexpected" }
