+++ state/Cargo.toml
@@ -23,5 +23,4 @@ rlp = { path = "../util/rlp" }
 rustc-hex = "1.0"
 serde = "1.0"
 serde_derive = "1.0"
-unexpected = { path = "../util/unexpected" }
 util-error = { path = "../util/error" }
