+++ state/src/impls/shard_level.rs
@@ -25,11 +25,11 @@ use ctypes::transaction::{
     AssetMintOutput, AssetTransferInput, AssetTransferOutput, Error as TransactionError, Outcome as TransactionOutcome,
     Transaction,
 };
+use ctypes::util::unexpected::Mismatch;
 use ctypes::{ShardId, WorldId};
 use cvm::{decode, execute, ScriptResult, VMConfig};
 use primitives::{Bytes, H256};
 use rlp::Encodable;
-use unexpected::Mismatch;
 
 use super::super::backend::{Backend, ShardBackend};
 use super::super::checkpoint::{CheckpointId, StateWithCheckpoint};
