+++ types/Cargo.toml
@@ -13,4 +13,3 @@ rlp = { path = "../util/rlp" }
 rlp_derive = { path = "../util/rlp_derive" }
 serde = "1.0"
 serde_derive = "1.0"
-unexpected = { path = "../util/unexpected" }
