+++ types/src/util/mod.rs
@@ -14,25 +14,4 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-extern crate byteorder;
-extern crate codechain_crypto as ccrypto;
-extern crate codechain_key as ckey;
-extern crate heapsize;
-extern crate primitives;
-#[cfg_attr(test, macro_use)]
-extern crate rlp;
-#[macro_use]
-extern crate rlp_derive;
-extern crate serde;
-#[macro_use]
-extern crate serde_derive;
-
-pub mod invoice;
-pub mod machine;
-pub mod parcel;
-pub mod transaction;
-pub mod util;
-
-pub type BlockNumber = u64;
-pub type ShardId = u16;
-pub type WorldId = u16;
+pub mod unexpected;
