--- util/unexpected/Cargo.toml
@@ -1,6 +1,0 @@
-[package]
-name = "unexpected"
-version = "0.1.0"
-authors = ["Parity Technologies <admin@parity.io>"]
-
-[dependencies]
